requirements.txt
Telethon
configparser
loguru
groq
python-dotenv